requirement.txt
streamlit
pandas
matplotlib
tensorflow==2.5.1
keras-nightly==2.5.0.dev2021032900
seaborn
numpy
scikit-learn

